Response to typos from PR and minor tweaks to get simple test running on iOS
