+++ Realm/Realm/Native/NativeCommon.cs
@@ -83,7 +83,7 @@ namespace Realms
                         var assemblyLocation = Path.GetDirectoryName((string)assemblyLocationPI.GetValue(typeof(NativeCommon).GetTypeInfo().Assembly));
                         var architecture = InteropConfig.Is64BitProcess ? "x64" : "x86";
                         var path = Path.Combine(assemblyLocation, "lib", "win32", architecture) + Path.PathSeparator + Environment.GetEnvironmentVariable("PATH");
-                         Environment.SetEnvironmentVariable("PATH", path);  // hack we can do for Windows but not UWP
+                        Environment.SetEnvironmentVariable("PATH", path);  // hack we can do for Windows but not WindowsStore environment for UWP, which includes the libs via <Content> tags
                     }
                 }
             }
