+++ Tests/Tests.XamarinIOS/UnitTestAppDelegate.cs
@@ -18,6 +18,7 @@
 
 using Foundation;
 using NUnit.Runner.Services;
+using NUnit.Tests.Simple;
 using UIKit;
 
 namespace IntegrationTests.XamarinIOS
@@ -31,7 +32,6 @@ namespace IntegrationTests.XamarinIOS
 
             // This will load all tests within the current project
             var runner = new NUnit.Runner.App();
-            runner.AddTestAssembly(System.Reflection.Assembly.GetExecutingAssembly());
 
             // Do you want to automatically run tests when the app starts?
             runner.Options = new TestOptions { AutoRun = false };
